added  pydantic settings for type-safe configuration and added type annotations throughout codebase
