+++ db_api/config/settings.py
@@ -1,41 +1,20 @@
 """
-Application settings configuration using Pydantic.
+Standard environment settings.
 
-This module loads environment variables using python-dotenv and defines a
-Pydantic settings class to hold configuration variables for the application.
+This module loads environment variables using python-dotenv and exposes
+the following module-level constants:
 
-Attributes:
     APP_DATABASE_URL (str): The URL for connecting to the database.
     JWT_SECRET (str): The secret key for encoding/decoding JWT tokens.
-    ACCESS_TOKEN_EXPIRE_MINUTES (int): Number of minutes before an access token expires.
+    ACCESS_TOKEN_EXPIRE_MINUTES (str): The access token expiration in minutes.
 """
 
 import os
 from dotenv import load_dotenv, find_dotenv
-from pydantic_settings import BaseSettings  # Updated import for Pydantic v2.5+
 
 # Load environment variables from a .env file.
 load_dotenv(find_dotenv())
 
-class Settings(BaseSettings):
-    """Pydantic settings for the application."""
-    app_database_url: str
-    jwt_secret: str
-    access_token_expire_minutes: int
-
-    class Config:
-        """
-        Pydantic configuration for the settings class.
-
-        The env_file is set to the location found by `find_dotenv()` (or defaults to ".env")
-        to ensure environment variables are loaded.
-        """
-        env_file = find_dotenv() or ".env"
-
-# Instantiate settings.
-settings: Settings = Settings()
-
-# Define module-level variables with type annotations.
-APP_DATABASE_URL: str = settings.app_database_url
-JWT_SECRET: str = settings.jwt_secret
-ACCESS_TOKEN_EXPIRE_MINUTES: int = settings.access_token_expire_minutes
+APP_DATABASE_URL = os.environ.get("APP_DATABASE_URL")
+JWT_SECRET = os.environ.get("JWT_SECRET")
+ACCESS_TOKEN_EXPIRE_MINUTES = os.environ.get("ACCESS_TOKEN_EXPIRE_MINUTES")
